requirements.txt
fastapi
uvicorn
torch
diffusers
pillow
transformers
accelerate